+++ custom_components/ipx800v4/__init__.py
@@ -1,4 +1,5 @@
 """Support for the GCE IPX800 V4."""
+from base64 import b64decode
 from datetime import timedelta
 import logging
 import re
@@ -37,10 +38,12 @@ from homeassistant.helpers.update_coordinator import (
 from .const import (
     CONF_COMPONENT,
     CONF_COMPONENT_ALLOWED,
+    CONF_DEFAULT_BRIGHTNESS,
     CONF_DEVICES,
     CONF_EXT_ID,
     CONF_ID,
     CONF_IDS,
+    CONF_PUSH_PASSWORD,
     CONF_TRANSITION,
     CONF_TYPE,
     CONF_TYPE_ALLOWED,
@@ -48,9 +51,11 @@ from .const import (
     COORDINATOR,
     DEFAULT_TRANSITION,
     DOMAIN,
+    PUSH_USERNAME,
     REQUEST_REFRESH_DELAY,
     TYPE_RELAY,
     TYPE_X4VR,
+    TYPE_XPWM,
     TYPE_XPWM_RGB,
     TYPE_XPWM_RGBW,
     UNDO_UPDATE_LISTENER,
@@ -66,6 +71,7 @@ DEVICE_CONFIG_SCHEMA_ENTRY = vol.Schema(
         vol.Optional(CONF_ID): cv.positive_int,
         vol.Optional(CONF_IDS): cv.ensure_list,
         vol.Optional(CONF_EXT_ID): cv.positive_int,
+        vol.Optional(CONF_DEFAULT_BRIGHTNESS): cv.positive_int,
         vol.Optional(CONF_ICON): cv.icon,
         vol.Optional(CONF_TRANSITION, default=DEFAULT_TRANSITION): vol.Coerce(float),
         vol.Optional(CONF_DEVICE_CLASS): cv.string,
@@ -100,7 +106,7 @@ async def async_setup(hass: HomeAssistantType, config: ConfigType) -> bool:
     hass.data.setdefault(DOMAIN, {})
 
     if DOMAIN in config:
-        for gateway in config.get(DOMAIN):
+        for gateway in config[DOMAIN]:
             hass.async_create_task(
                 hass.config_entries.flow.async_init(
                     DOMAIN, context={"source": SOURCE_IMPORT}, data=gateway
@@ -144,7 +150,7 @@ async def async_setup_entry(hass: HomeAssistantType, entry: ConfigEntry) -> bool
         except Ipx800CannotConnectError as err:
             raise UpdateFailed(f"Failed to communicating with API: {err}") from err
 
-    scan_interval = int(entry.data.get(CONF_SCAN_INTERVAL))
+    scan_interval = entry.data[CONF_SCAN_INTERVAL]
 
     if scan_interval < 10:
         _LOGGER.warning(
@@ -206,8 +212,18 @@ async def async_setup_entry(hass: HomeAssistantType, entry: ConfigEntry) -> bool
         )
 
     # Provide endpoints for the IPX to call to push states
-    hass.http.register_view(IpxRequestView)
-    hass.http.register_view(IpxRequestDataView)
+    if CONF_PUSH_PASSWORD in entry.data:
+        hass.http.register_view(
+            IpxRequestView(entry.data[CONF_HOST], entry.data[CONF_PUSH_PASSWORD])
+        )
+        hass.http.register_view(
+            IpxRequestDataView(entry.data[CONF_HOST], entry.data[CONF_PUSH_PASSWORD])
+        )
+    else:
+        _LOGGER.info(
+            "No %s parameter provided in configuration, skip API call handling for IPX800 PUSH.",
+            CONF_PUSH_PASSWORD,
+        )
 
     return True
 
@@ -265,6 +281,31 @@ def build_device_list(devices_config: list) -> list:
             )
             continue
 
+        # Check if only PWM have default_brightness set
+        if CONF_DEFAULT_BRIGHTNESS in device_config and not (
+            device_config[CONF_TYPE] == TYPE_XPWM
+            or device_config[CONF_TYPE] == TYPE_XPWM_RGB
+            or device_config[CONF_TYPE] == TYPE_XPWM_RGBW
+        ):
+            _LOGGER.error(
+                "Device %s skipped: %s must be set only for XPWM types.",
+                device_config[CONF_NAME],
+                CONF_DEFAULT_BRIGHTNESS,
+            )
+            continue
+
+        # Check if only PWM have default_brightness set
+        if CONF_DEFAULT_BRIGHTNESS in device_config and (
+            device_config[CONF_DEFAULT_BRIGHTNESS] < 1
+            or device_config[CONF_DEFAULT_BRIGHTNESS] > 255
+        ):
+            _LOGGER.error(
+                "Device %s skipped: %s must be between 1 and 255.",
+                device_config[CONF_NAME],
+                CONF_DEFAULT_BRIGHTNESS,
+            )
+            continue
+
         # Check if RGB/RBW or FP/RELAY have ids set
         if (
             device_config[CONF_TYPE] == TYPE_XPWM_RGB
@@ -319,8 +360,25 @@ class IpxRequestView(HomeAssistantView):
     url = "/api/ipx800v4/{entity_id}/{state}"
     name = "api:ipx800v4"
 
+    def __init__(self, host: str, password: str) -> None:
+        """Init the IPX view."""
+        self.host = host
+        self.password = password
+        super().__init__()
+
     async def get(self, request, entity_id, state):
         """Respond to requests from the device."""
+        if request.remote != self.host:
+            raise ApiCallNotAuthorized("API call not coming from IPX800 IP.")
+        if "Authorization" not in request.headers:
+            raise ApiCallNotAuthorized("API call no authentication provided.")
+        header_auth = request.headers["Authorization"]
+        split = header_auth.strip().split(" ")
+        if len(split) != 2 or split[0].strip().lower() != "basic":
+            raise ApiCallNotAuthorized("Malformed Authorization header")
+        username, password = b64decode(split[1]).decode().split(":", 1)
+        if username != PUSH_USERNAME and password != self.password:
+            raise ApiCallNotAuthorized("API call authentication invalid.")
         hass = request.app["hass"]
         old_state = hass.states.get(entity_id)
         _LOGGER.debug("Update %s to state %s.", entity_id, state)
@@ -337,8 +395,25 @@ class IpxRequestDataView(HomeAssistantView):
     url = "/api/ipx800v4_data/{data}"
     name = "api:ipx800v4_data"
 
+    def __init__(self, host: str, password: str) -> None:
+        """Init the IPX view."""
+        self.host = host
+        self.password = password
+        super().__init__()
+
     async def get(self, request, data):
         """Respond to requests from the device."""
+        if request.remote != self.host:
+            raise ApiCallNotAuthorized("API call not coming from IPX800 IP.")
+        if "Authorization" not in request.headers:
+            raise ApiCallNotAuthorized("API call no authentication provided.")
+        header_auth = request.headers["Authorization"]
+        split = header_auth.strip().split(" ")
+        if len(split) != 2 or split[0].strip().lower() != "basic":
+            raise ApiCallNotAuthorized("Malformed Authorization header")
+        username, password = b64decode(split[1]).decode().split(":", 1)
+        if username != PUSH_USERNAME and password != self.password:
+            raise ApiCallNotAuthorized("API call authentication invalid.")
         hass = request.app["hass"]
         entities_data = data.split("&")
         for entity_data in entities_data:
@@ -353,6 +428,10 @@ class IpxRequestDataView(HomeAssistantView):
                 _LOGGER.warning("Entity not found for state updating: %s", entity_id)
 
         return web.Response(status=HTTP_OK, text="OK")
+
+
+class ApiCallNotAuthorized(BaseException):
+    """API call for IPX800 view not authorized."""
 
 
 class IpxDevice(CoordinatorEntity):
@@ -373,7 +452,7 @@ class IpxDevice(CoordinatorEntity):
         self._name = device_config.get(CONF_NAME)
         self._device_name = self._name
         if suffix_name:
-            self._name += f" {suffix_name}"
+            self._name = f"{self._name} {suffix_name}"
 
         self._device_class = device_config.get(CONF_DEVICE_CLASS)
         self._unit_of_measurement = device_config.get(CONF_UNIT_OF_MEASUREMENT)