+++ custom_components/ipx800v4/const.py
@@ -5,6 +5,7 @@ CONTROLLER = "controller"
 COORDINATOR = "coordinator"
 UNDO_UPDATE_LISTENER = "undo_update_listener"
 GLOBAL_PARALLEL_UPDATES = 1
+PUSH_USERNAME = "ipx800"
 
 DEFAULT_TRANSITION = 0.5
 REQUEST_REFRESH_DELAY = 0.5
@@ -12,9 +13,11 @@ REQUEST_REFRESH_DELAY = 0.5
 CONF_DEVICES = "devices"
 
 CONF_COMPONENT = "component"
+CONF_DEFAULT_BRIGHTNESS = "default_brightness"
 CONF_ID = "id"
 CONF_IDS = "ids"
 CONF_EXT_ID = "ext_id"
+CONF_PUSH_PASSWORD = "push_password"
 CONF_TRANSITION = "transition"
 CONF_TYPE = "type"
 
@@ -26,6 +29,7 @@ TYPE_XDIMMER = "xdimmer"
 TYPE_VIRTUALOUT = "virtualout"
 TYPE_VIRTUALIN = "virtualin"
 TYPE_ANALOGIN = "analogin"
+TYPE_VIRTUALANALOGIN = "virtualanalogin"
 TYPE_DIGITALIN = "digitalin"
 TYPE_X4VR = "x4vr"
 TYPE_XTHL = "xthl"
@@ -49,6 +53,7 @@ CONF_TYPE_ALLOWED = [
     TYPE_VIRTUALOUT,
     TYPE_VIRTUALIN,
     TYPE_ANALOGIN,
+    TYPE_VIRTUALANALOGIN,
     TYPE_DIGITALIN,
     TYPE_X4VR,
     TYPE_XTHL,