+++ custom_components/ipx800v4/light.py
@@ -1,27 +1,28 @@
 """Support for IPX800 V4 lights."""
 import logging
+from typing import List
 
 from pypx800 import IPX800, XPWM, Ipx800RequestError, Relay, XDimmer
 
 from homeassistant.components.light import (
     ATTR_BRIGHTNESS,
-    ATTR_HS_COLOR,
     ATTR_RGB_COLOR,
+    ATTR_RGBW_COLOR,
     ATTR_TRANSITION,
-    ATTR_WHITE_VALUE,
-    SUPPORT_BRIGHTNESS,
-    SUPPORT_COLOR,
+    COLOR_MODE_BRIGHTNESS,
+    COLOR_MODE_ONOFF,
+    COLOR_MODE_RGB,
+    COLOR_MODE_RGBW,
     SUPPORT_TRANSITION,
-    SUPPORT_WHITE_VALUE,
     LightEntity,
 )
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
-import homeassistant.util.color as color_util
 
 from . import IpxDevice
 from .const import (
+    CONF_DEFAULT_BRIGHTNESS,
     CONF_DEVICES,
     CONF_TRANSITION,
     CONF_TYPE,
@@ -61,7 +62,7 @@ async def async_setup_entry(
     coordinator = hass.data[DOMAIN][entry.entry_id][COORDINATOR]
     devices = hass.data[DOMAIN][entry.entry_id][CONF_DEVICES]["light"]
 
-    entities = []
+    entities: List[LightEntity] = []
 
     for device in devices:
         if device.get(CONF_TYPE) == TYPE_RELAY:
@@ -90,6 +91,16 @@ class RelayLight(IpxDevice, LightEntity):
         """Initialize the RelayLight."""
         super().__init__(device_config, ipx, coordinator)
         self.control = Relay(ipx, self._id)
+
+    @property
+    def supported_color_modes(self) -> set:
+        """Return supported color modes."""
+        return {COLOR_MODE_ONOFF}
+
+    @property
+    def color_mode(self) -> str:
+        """Return the color mode of the light."""
+        return COLOR_MODE_ONOFF
 
     @property
     def is_on(self) -> bool:
@@ -144,9 +155,19 @@ class XDimmerLight(IpxDevice, LightEntity):
         self._transition = device_config.get(CONF_TRANSITION, DEFAULT_TRANSITION)
 
     @property
-    def supported_features(self):
-        """Flag supported features."""
-        return SUPPORT_BRIGHTNESS | SUPPORT_TRANSITION
+    def supported_color_modes(self) -> set:
+        """Return supported color modes."""
+        return {COLOR_MODE_BRIGHTNESS}
+
+    @property
+    def supported_features(self) -> int:
+        """Return supported features."""
+        return SUPPORT_TRANSITION
+
+    @property
+    def color_mode(self) -> str:
+        """Return the color mode of the light."""
+        return COLOR_MODE_BRIGHTNESS
 
     @property
     def is_on(self) -> bool:
@@ -162,9 +183,9 @@ class XDimmerLight(IpxDevice, LightEntity):
         """Turn on the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
             if ATTR_BRIGHTNESS in kwargs:
-                self._brightness = kwargs.get(ATTR_BRIGHTNESS, 255)
+                self._brightness = kwargs[ATTR_BRIGHTNESS]
                 await self.control.set_level(
                     scaleto100(self._brightness), self._transition * 1000
                 )
@@ -180,7 +201,7 @@ class XDimmerLight(IpxDevice, LightEntity):
         """Turn off the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
             await self.control.off(self._transition * 1000)
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
@@ -192,7 +213,7 @@ class XDimmerLight(IpxDevice, LightEntity):
         """Toggle the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
             await self.control.toggle(self._transition * 1000)
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
@@ -213,12 +234,23 @@ class XPWMLight(IpxDevice, LightEntity):
         self.control = XPWM(ipx, self._id)
 
         self._brightness = None
+        self._default_brightness = device_config.get(CONF_DEFAULT_BRIGHTNESS, 100)
         self._transition = device_config.get(CONF_TRANSITION, DEFAULT_TRANSITION)
 
     @property
-    def supported_features(self):
-        """Flag supported features."""
-        return SUPPORT_BRIGHTNESS | SUPPORT_TRANSITION
+    def supported_color_modes(self) -> set:
+        """Return supported color modes."""
+        return {COLOR_MODE_BRIGHTNESS}
+
+    @property
+    def supported_features(self) -> int:
+        """Return supported features."""
+        return SUPPORT_TRANSITION
+
+    @property
+    def color_mode(self) -> str:
+        """Return the color mode of the light."""
+        return COLOR_MODE_BRIGHTNESS
 
     @property
     def is_on(self) -> bool:
@@ -234,14 +266,16 @@ class XPWMLight(IpxDevice, LightEntity):
         """Turn on the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
             if ATTR_BRIGHTNESS in kwargs:
-                self._brightness = kwargs.get(ATTR_BRIGHTNESS, 255)
+                self._brightness = kwargs[ATTR_BRIGHTNESS]
                 await self.control.set_level(
                     scaleto100(self._brightness), self._transition * 1000
                 )
             else:
-                await self.control.on(self._transition * 1000)
+                await self.control.set_level(
+                    self._default_brightness, self._transition * 1000
+                )
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
             _LOGGER.error(
@@ -252,7 +286,7 @@ class XPWMLight(IpxDevice, LightEntity):
         """Turn off the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
             await self.control.off(self._transition * 1000)
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
@@ -264,7 +298,7 @@ class XPWMLight(IpxDevice, LightEntity):
         """Toggle the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
             await self.control.toggle(self._transition * 1000)
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
@@ -287,29 +321,33 @@ class XPWMRGBLight(IpxDevice, LightEntity):
         self.xpwm_rgb_b = XPWM(ipx, self._ids[2])
 
         self._brightness = None
-        self._rgb_color = None
+        self._default_brightness = device_config.get(CONF_DEFAULT_BRIGHTNESS, 100)
         self._transition = device_config.get(CONF_TRANSITION, DEFAULT_TRANSITION)
 
     @property
-    def supported_features(self):
-        """Flag supported features."""
-        return SUPPORT_BRIGHTNESS | SUPPORT_TRANSITION | SUPPORT_COLOR
+    def supported_color_modes(self) -> set:
+        """Return supported color modes."""
+        return {COLOR_MODE_RGB}
+
+    @property
+    def supported_features(self) -> int:
+        """Return supported features."""
+        return SUPPORT_TRANSITION
+
+    @property
+    def color_mode(self) -> str:
+        """Return the color mode of the light."""
+        return COLOR_MODE_RGB
 
     @property
     def is_on(self) -> bool:
         """Return if at least a level in on."""
-        level_r = scaleto255(self.coordinator.data[f"PWM{self._ids[0]}"])
-        level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
-        level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
-        return level_r > 0 or level_b > 0 or level_g > 0
+        return any(i > 0 for i in self.rgb_color)
 
     @property
-    def brightness(self):
+    def brightness(self) -> int:
         """Return the brightness from levels."""
-        level_r = scaleto255(self.coordinator.data[f"PWM{self._ids[0]}"])
-        level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
-        level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
-        return 0.2126 * level_r + 0.7152 * level_g + 0.0722 * level_b
+        return max(self.rgb_color)
 
     @property
     def rgb_color(self):
@@ -317,47 +355,43 @@ class XPWMRGBLight(IpxDevice, LightEntity):
         level_r = scaleto255(self.coordinator.data[f"PWM{self._ids[0]}"])
         level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
         level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
-        return [level_r, level_g, level_b]
+        return (level_r, level_g, level_b)
 
     async def async_turn_on(self, **kwargs) -> None:
         """Turn on the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
-
-            if ATTR_RGB_COLOR in kwargs or ATTR_HS_COLOR in kwargs:
-                if ATTR_RGB_COLOR in kwargs:
-                    self._rgb_color = kwargs.get(ATTR_RGB_COLOR)
-                elif ATTR_HS_COLOR in kwargs:
-                    self._rgb_color = color_util.color_hs_to_RGB(
-                        *kwargs.get(ATTR_HS_COLOR)
-                    )
+                self._transition = kwargs[ATTR_TRANSITION]
+            if ATTR_RGB_COLOR in kwargs:
+                colors = kwargs[ATTR_RGB_COLOR]
                 await self.xpwm_rgb_r.set_level(
-                    scaleto100(self._rgb_color[0]),
-                    self._transition * 1000,
+                    scaleto100(colors[0]), self._transition * 1000
                 )
                 await self.xpwm_rgb_g.set_level(
-                    scaleto100(self._rgb_color[1]),
-                    self._transition * 1000,
+                    scaleto100(colors[1]), self._transition * 1000
                 )
                 await self.xpwm_rgb_b.set_level(
-                    scaleto100(self._rgb_color[2]),
-                    self._transition * 1000,
+                    scaleto100(colors[2]), self._transition * 1000
                 )
-
             elif ATTR_BRIGHTNESS in kwargs:
-                self._brightness = kwargs.get(ATTR_BRIGHTNESS)
+                self._brightness = kwargs[ATTR_BRIGHTNESS]
                 if self.is_on:
                     await self.xpwm_rgb_r.set_level(
-                        scaleto100(self.rgb_color[0] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[0]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                     await self.xpwm_rgb_g.set_level(
-                        scaleto100(self.rgb_color[1] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[1]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                     await self.xpwm_rgb_b.set_level(
-                        scaleto100(self.rgb_color[2] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[2]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                 else:
@@ -374,18 +408,26 @@ class XPWMRGBLight(IpxDevice, LightEntity):
                         self._transition * 1000,
                     )
             else:
-                await self.xpwm_rgb_r.on(self._transition * 1000)
-                await self.xpwm_rgb_g.on(self._transition * 1000)
-                await self.xpwm_rgb_b.on(self._transition * 1000)
+                await self.xpwm_rgb_r.set_level(
+                    self._default_brightness, self._transition * 1000
+                )
+                await self.xpwm_rgb_g.set_level(
+                    self._default_brightness, self._transition * 1000
+                )
+                await self.xpwm_rgb_b.set_level(
+                    self._default_brightness, self._transition * 1000
+                )
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
-            _LOGGER.error("An error occurred while turn on IPX800 light: %s", self.name)
+            _LOGGER.error(
+                "An error occurred while turn off IPX800 light: %s", self.name
+            )
 
     async def async_turn_off(self, **kwargs) -> None:
         """Turn off the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION) * 1000)
+                self._transition = kwargs[ATTR_TRANSITION] * 1000
             await self.xpwm_rgb_r.off(self._transition * 1000)
             await self.xpwm_rgb_g.off(self._transition * 1000)
             await self.xpwm_rgb_b.off(self._transition * 1000)
@@ -413,49 +455,33 @@ class XPWMRGBWLight(IpxDevice, LightEntity):
         self.xpwm_rgbw_w = XPWM(ipx, self._ids[3])
 
         self._brightness = None
-        self._rgb_color = None
-        self._white_value = None
+        self._default_brightness = device_config.get(CONF_DEFAULT_BRIGHTNESS, 100)
         self._transition = device_config.get(CONF_TRANSITION, DEFAULT_TRANSITION)
 
     @property
-    def supported_features(self):
-        """Flag supported features."""
-        return (
-            SUPPORT_BRIGHTNESS
-            | SUPPORT_TRANSITION
-            | SUPPORT_COLOR
-            | SUPPORT_WHITE_VALUE
-        )
+    def supported_color_modes(self) -> set:
+        """Return supported color modes."""
+        return {COLOR_MODE_RGBW}
+
+    @property
+    def supported_features(self) -> int:
+        """Return supported features."""
+        return SUPPORT_TRANSITION
+
+    @property
+    def color_mode(self) -> str:
+        """Return the color mode of the light."""
+        return COLOR_MODE_RGBW
 
     @property
     def is_on(self) -> bool:
         """Return if at least a level in on."""
-        level_r = scaleto255(self.coordinator.data[f"PWM{self._ids[0]}"])
-        level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
-        level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
-        level_w = scaleto255(self.coordinator.data[f"PWM{self._ids[3]}"])
-        return level_r > 0 or level_b > 0 or level_g > 0 or level_w > 0
+        return any(i > 0 for i in self.rgbw_color)
 
     @property
-    def brightness(self):
+    def brightness(self) -> int:
         """Return the brightness from levels."""
-        level_r = scaleto255(self.coordinator.data[f"PWM{self._ids[0]}"])
-        level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
-        level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
-        level_w = scaleto255(self.coordinator.data[f"PWM{self._ids[3]}"])
-        if level_r > 0 or level_b > 0 or level_g > 0:
-            self._brightness = 0.2126 * level_r + 0.7152 * level_g + 0.0722 * level_b
-        else:
-            self._brightness = level_w
-        return self._brightness
-
-    @property
-    def rgb_color(self):
-        """Return the RGB color from RGB levels."""
-        level_r = scaleto255(self.coordinator.data[f"PWM{self._ids[0]}"])
-        level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
-        level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
-        return [level_r, level_g, level_b]
+        return max(self.rgbw_color)
 
     @property
     def rgbw_color(self):
@@ -464,61 +490,54 @@ class XPWMRGBWLight(IpxDevice, LightEntity):
         level_g = scaleto255(self.coordinator.data[f"PWM{self._ids[1]}"])
         level_b = scaleto255(self.coordinator.data[f"PWM{self._ids[2]}"])
         level_w = scaleto255(self.coordinator.data[f"PWM{self._ids[3]}"])
-        return [level_r, level_g, level_b, level_w]
-
-    @property
-    def white_value(self):
-        """Return the white value from white level."""
-        return scaleto255(self.coordinator.data[f"PWM{self._ids[3]}"])
+        return (level_r, level_g, level_b, level_w)
 
     async def async_turn_on(self, **kwargs) -> None:
         """Turn on the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
 
-            if ATTR_RGB_COLOR in kwargs or ATTR_HS_COLOR in kwargs:
-                if ATTR_RGB_COLOR in kwargs:
-                    self._rgb_color = kwargs.get(ATTR_RGB_COLOR)
-                elif ATTR_HS_COLOR in kwargs:
-                    self._rgb_color = color_util.color_hs_to_RGB(
-                        *kwargs.get(ATTR_HS_COLOR)
-                    )
-                rgbw = color_util.color_rgb_to_rgbw(*self._rgb_color)
+            if ATTR_RGBW_COLOR in kwargs:
+                colors = kwargs[ATTR_RGBW_COLOR]
+                # if only rgb color have been set
                 await self.xpwm_rgbw_r.set_level(
-                    scaleto100(rgbw[0]), self._transition * 1000
+                    scaleto100(colors[0]), self._transition * 1000
                 )
                 await self.xpwm_rgbw_g.set_level(
-                    scaleto100(rgbw[1]), self._transition * 1000
+                    scaleto100(colors[1]), self._transition * 1000
                 )
                 await self.xpwm_rgbw_b.set_level(
-                    scaleto100(rgbw[2]), self._transition * 1000
+                    scaleto100(colors[2]), self._transition * 1000
                 )
                 await self.xpwm_rgbw_w.set_level(
-                    scaleto100(rgbw[3]), self._transition * 1000
-                )
-            elif ATTR_WHITE_VALUE in kwargs:
-                self._white_value = kwargs.get(ATTR_WHITE_VALUE)
-                await self.xpwm_rgbw_w.set_level(
-                    scaleto100(self._white_value), self._transition * 1000
+                    scaleto100(colors[3]), self._transition * 1000
                 )
             elif ATTR_BRIGHTNESS in kwargs:
-                self._brightness = kwargs.get(ATTR_BRIGHTNESS)
+                self._brightness = kwargs[ATTR_BRIGHTNESS]
                 if self.is_on:
                     await self.xpwm_rgbw_r.set_level(
-                        scaleto100(self.rgbw_color[0] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[0]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                     await self.xpwm_rgbw_g.set_level(
-                        scaleto100(self.rgbw_color[1] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[1]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                     await self.xpwm_rgbw_b.set_level(
-                        scaleto100(self.rgbw_color[2] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[2]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                     await self.xpwm_rgbw_w.set_level(
-                        scaleto100(self.rgbw_color[3] * self._brightness / 255),
+                        self.coordinator.data[f"PWM{self._ids[3]}"]
+                        * self._brightness
+                        / 100,
                         self._transition * 1000,
                     )
                 else:
@@ -527,7 +546,9 @@ class XPWMRGBWLight(IpxDevice, LightEntity):
                         self._transition * 1000,
                     )
             else:
-                await self.xpwm_rgbw_w.on(self._transition * 1000)
+                await self.xpwm_rgbw_w.set_level(
+                    self._default_brightness, self._transition * 1000
+                )
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
             _LOGGER.error(
@@ -538,11 +559,11 @@ class XPWMRGBWLight(IpxDevice, LightEntity):
         """Turn off the light."""
         try:
             if ATTR_TRANSITION in kwargs:
-                self._transition = int(kwargs.get(ATTR_TRANSITION))
+                self._transition = kwargs[ATTR_TRANSITION]
+            await self.xpwm_rgbw_w.off(self._transition * 1000)
             await self.xpwm_rgbw_r.off(self._transition * 1000)
             await self.xpwm_rgbw_g.off(self._transition * 1000)
             await self.xpwm_rgbw_b.off(self._transition * 1000)
-            await self.xpwm_rgbw_w.off(self._transition * 1000)
             await self.coordinator.async_request_refresh()
         except Ipx800RequestError:
             _LOGGER.error(